implement logic and fill out test methods + renaming
